+++ projects/portal/src/app/app-routing.module.ts
@@ -1,8 +1,6 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
-
 import { LoginComponent } from './pages/login/login-portal-paciente';
-import { PortalPacienteComponent } from './pages/portal-paciente/portal-paciente.component';
 import { RoutingGuard } from './app.routings-guard.class';
 import { PDPMisFamiliaresComponent } from './pages/mis-familiares/mis-familiares.component';
 import { PDPMisFamiliaresDetalleComponent } from './pages/mis-familiares/mis-familiares-detalle.component';
@@ -10,7 +8,6 @@ import { PDPMisFamiliaresDetalleComponent } from './pages/mis-familiares/mis-fam
 const routes: Routes = [
     { path: '', redirectTo: '/login', pathMatch: 'full' },
     { path: 'login', component: LoginComponent },
-    { path: 'home', component: PortalPacienteComponent, canActivate: [RoutingGuard] },
     {
         path: 'mis-familiares',
         component: PDPMisFamiliaresComponent,
