+++ projects/portal/src/app/components/portal-menu/portal-menu.component.html
@@ -5,8 +5,8 @@
                size="xs">
         <ng-container *ngFor="let card of cards">
             <plex-label icon="{{ card.icono }}" size="md" direction="column" class="mb-2 color-{{ card.semanticTag}}"
-                        titulo="" subtitulo="" [routerLink]="[{ outlets: { listado: [card.path] } }]"
-                        tooltip="{{ card.path }}" tooltipPosition="right">
+                        titulo="" subtitulo="" (click)="gotTo(card.path)" tooltip="{{ card.path }}"
+                        tooltipPosition="right">
             </plex-label>
         </ng-container>
     </plex-grid>
