--- projects/portal/src/app/mock-data/mock-cards.ts
@@ -1,14 +1,0 @@
-import { ICard } from '../interfaces/icard';
-
-export const CARDS: ICard[] = [
-    {
-        id: 10,
-        nombre: 'relaciones',
-        tipo: 'info',
-        semanticTag: 'solicitud',
-        icono: 'familia',
-        path: 'misFamiliares',
-        color: '#0070cc',
-        outlet: 'listado'
-    }
-];
