--- projects/portal/src/app/pages/portal-paciente/portal-paciente-main/portal-paciente-main.html
@@ -1,14 +1,0 @@
-<pdp-paciente-detalle></pdp-paciente-detalle>
-<hr>
-<plex-title size="sm" titulo="mis consultas">
-    <plex-badge *mobile="false" type="info">{{ hoy | fecha }}</plex-badge>
-    <plex-button type="success" icon="plus" tooltip="agregar registro" class="ml-2" size="sm">
-    </plex-button>
-
-    <plex-button type="info" icon="zoom" tooltip="mostrar filtros" class="ml-2" size="sm" (click)="enviarFiltros()">
-    </plex-button>
-</plex-title>
-
-
-
-<router-outlet name="listado"></router-outlet>
