--- projects/portal/src/app/pages/portal-paciente/portal-paciente-sidebar/detalle-familiar.html
@@ -1,24 +1,0 @@
-<ng-container *ngIf="familiar">
-    <plex-detail size="lg">
-        <img [mpiFotoPaciente]="familiar">
-        <plex-badge type="success">{{ vinculoFamiliar }}</plex-badge>
-        <div title>{{ familiar.apellido }}, {{ familiar.nombre }} </div>
-        <div subtitle>{{ familiar.documento }} </div>
-        <plex-label titulo="Edad" subtitulo="{{ familiar.edad }}"></plex-label>
-        <plex-label titulo="Documento" subtitulo="{{ familiar.documento }}"></plex-label>
-    </plex-detail>
-    <plex-title titulo="datos de contacto" size="sm"></plex-title>
-    <plex-grid type="auto" size="md">
-        <plex-label icon="celular" titulo="telefono" subtitulo="{{ familiar.contacto[0]?.valor || 'Sin datos'}}">
-        </plex-label>
-        <plex-label icon="home" titulo="domicilio" subtitulo="{{ familiar.direccion[0]?.valor || 'Sin datos'}}">
-        </plex-label>
-        <plex-label icon="city" titulo="localidad"
-                    subtitulo="{{ familiar.direccion[0]?.ubicacion?.localidad?.nombre || 'Sin datos'}}"></plex-label>
-        <plex-label icon="map-marker" titulo="provincia"
-                    subtitulo="{{ familiar.direccion[0]?.ubicacion?.provincia?.nombre || 'Sin datos'}}"></plex-label>
-    </plex-grid>
-    <hr>
-    <plex-button class="mt-4" *ngIf="vinculoFamiliar === 'hijo/a'" type="warning" size="block" label="acceder a HUDS">
-    </plex-button>
-</ng-container>
