--- projects/portal/src/app/pages/portal-paciente/portal-paciente.html
@@ -1,65 +1,0 @@
-<plex-layout main="{{ sidebarValue }}" resizable="false" foco="{{ valorFoco }}">
-    <plex-layout-main type="invert">
-        <plex-title size="{{ isResponsive() ? 'sm' : 'md' }}"
-                    titulo="{{ isResponsive() ? 'mi portal' : 'mi portal de salud' }}" main>
-            <plex-button size="sm" [icon]="isResponsive() ? 'mail-outline' : null" class="ml-2" type="link">
-                info@andes.gob.ar
-            </plex-button>
-            <plex-button size="sm" [label]="isResponsive() ? null : 'www.andes.gob.ar'" class="ml-2" type="link">
-                <ng-container>
-                    <plex-icon name="web"></plex-icon>
-                </ng-container>
-            </plex-button>
-            <plex-button type="link" icon="message-alert" size="sm" hint="tienes dos elementos sin leer" hintIcon="bell"
-                         hintType="warning" detach="top" [routerLink]="[{ outlets: { listado: ['misMensajes'] } }]">
-            </plex-button>
-            <plex-button *mobile="false" size="sm" class="ml-2" type="danger" label="cerrar sesión" routerLink="/login">
-            </plex-button>
-        </plex-title>
-
-
-        <plex-grid cols="12" size="md">
-
-            <!-- Menu Lateral -->
-            <plex-grid id="menu-vertical" *ngIf="!isResponsive()" span="{{ valorMenu }}"
-                       class="d-flex align-self-start mr-2 flex-column" size="xs">
-                <plex-button *ngIf="valorMenu === 2" size="sm" type="link" icon="close" (click)="contraerMenu()">
-                </plex-button>
-                <ng-container *ngFor="let card of cards$ | async">
-                    <plex-label icon="{{ card.icono }}" size="md" direction="column"
-                                class="mb-2 color-{{ card.semanticTag}}" titulo="" subtitulo=""
-                                [routerLink]="[{ outlets: { listado: [card.path] } }]" tooltip="{{ card.path }}"
-                                tooltipPosition="right">
-                    </plex-label>
-                </ng-container>
-            </plex-grid>
-
-            <!-- Listado -->
-            <section span="{{ isResponsive() ? '12' : '11' }}">
-                <pdp-portal-paciente-main (eventoSidebar)="recibirSidebar($event)" (eventoFoco)="recibirFoco($event)">
-                </pdp-portal-paciente-main>
-            </section>
-        </plex-grid>
-
-    </plex-layout-main>
-
-    <plex-layout-sidebar>
-        <plex-title size="{{ isResponsive() ? 'md' : 'sm' }}"
-                    titulo="{{ isResponsive() ? 'DETALLE' : 'detalle de la consulta' }}">
-            <plex-button type="danger" icon="close" size="sm" (click)="contraerSidebar()" class="ml-2"></plex-button>
-        </plex-title>
-        <!--<router-outlet name="detalle"></router-outlet>-->
-        <router-outlet name="detalleHuds"></router-outlet>
-        <router-outlet name="detallePrescripcion"></router-outlet>
-        <router-outlet name="detalleVacuna"></router-outlet>
-        <router-outlet name="detalleLaboratorio"></router-outlet>
-        <router-outlet name="detalleTurno"></router-outlet>
-        <router-outlet name="detalleFamiliar"></router-outlet>
-        <router-outlet name="detalleProblema"></router-outlet>
-        <router-outlet name="detalleConsulta"></router-outlet>
-        <router-outlet name="detalleProfesional"></router-outlet>
-        <router-outlet name="detalleMensaje"></router-outlet>
-        <router-outlet name="detalleOrganizacion"></router-outlet>
-        <router-outlet name="detalleSolicitud"></router-outlet>
-    </plex-layout-sidebar>
-</plex-layout>
