--- projects/portal/src/app/pages/portal-paciente/portal-paciente.scss
@@ -1,10 +1,0 @@
-// Override Modal
-.font-title {
-    font-size: 1.8rem !important;
-}
-.font-subtitle {
-    font-size: 1.3rem !important;
-}
-.label-black {
-    color: black !important;
-}
